Fix milling from nodejs applications (#4)
* Ported over work from https://github.com/robbynshaw/textile-js-sdk

In browser file drop works with the client currently.

* Added usage to the README. Deleted and rebuilt docs with new name in package.json.

Renamed all the modules to better match the naming conventions in textile-go.
Fixed the url for better resiliency.

* Added the ability to pass FormData via a function

Fixed a bug in the core/api.js whereby headers were getting
duplicated and/or overwritten.

In the previous implementation, streams did not work because
milling requires the stream to be re-opened. Converting the
FormData creation into a function call alleviates this issue.